+++ src/core/distributions/math/DecomposedInverseWishartDistribution.cpp
@@ -8,7 +8,7 @@
 
 using namespace RevBayesCore;
 
-DecomposedInverseWishartDistribution::DecomposedInverseWishartDistribution(const TypedDagNode<MatrixReal> *insigma0, const TypedDagNode<int>* indf)  :
+DecomposedInverseWishartDistribution::DecomposedInverseWishartDistribution(const TypedDagNode<MatrixReal> *insigma0, const TypedDagNode<long>* indf)  :
 TypedDistribution<RevBayesCore::MatrixReal>(new MatrixReal(insigma0->getValue().getDim())),
 sigma0(insigma0),
 kappaVector(NULL),
@@ -25,7 +25,7 @@ dim( NULL )  {
     redrawValue();
 }
 
-DecomposedInverseWishartDistribution::DecomposedInverseWishartDistribution(const TypedDagNode<RbVector<double> > *inkappaVector, const TypedDagNode<int>* indf)  :
+DecomposedInverseWishartDistribution::DecomposedInverseWishartDistribution(const TypedDagNode<RbVector<double> > *inkappaVector, const TypedDagNode<long>* indf)  :
 TypedDistribution<RevBayesCore::MatrixReal>(new MatrixReal( inkappaVector->getValue().size()) ),
     sigma0(NULL),
     kappaVector(inkappaVector),
@@ -42,7 +42,7 @@ TypedDistribution<RevBayesCore::MatrixReal>(new MatrixReal( inkappaVector->getVa
     redrawValue();
 }
 
-DecomposedInverseWishartDistribution::DecomposedInverseWishartDistribution(const TypedDagNode<int>* indim, const TypedDagNode<double> *inkappa, const TypedDagNode<int>* indf)  :
+DecomposedInverseWishartDistribution::DecomposedInverseWishartDistribution(const TypedDagNode<long>* indim, const TypedDagNode<double> *inkappa, const TypedDagNode<long>* indf)  :
 TypedDistribution<RevBayesCore::MatrixReal>(new MatrixReal( size_t(indim->getValue()) )),
     sigma0(NULL),
     kappaVector(NULL),
@@ -81,11 +81,11 @@ void DecomposedInverseWishartDistribution::swapParameterInternal(const DagNode *
         }
     if (oldP == dim)
         {
-        dim = static_cast<const TypedDagNode<int>* >(newP);
+        dim = static_cast<const TypedDagNode<long>* >(newP);
         }
     if (oldP == df)
         {
-        df = static_cast<const TypedDagNode<int>* >(newP);
+        df = static_cast<const TypedDagNode<long>* >(newP);
         }
 }
