+++ src/core/distributions/math/StudentTDistribution.cpp
@@ -14,7 +14,7 @@
 using namespace RevBayesCore;
 
 
-StudentTDistribution::StudentTDistribution(const TypedDagNode<int> *df) : ContinuousDistribution( new double( 0.0 ) ),
+StudentTDistribution::StudentTDistribution(const TypedDagNode<long> *df) : ContinuousDistribution( new double( 0.0 ) ),
 degrees( df )
 {
     // add the parameters to our set (in the base class)
@@ -88,7 +88,7 @@ void StudentTDistribution::swapParameterInternal(const DagNode *oldP, const DagN
     
     if (oldP == degrees)
     {
-        degrees = static_cast<const TypedDagNode<int>* >( newP );
+        degrees = static_cast<const TypedDagNode<long>* >( newP );
     }
     
 }
